trivial_numeric_casts

+++ components/calendar/src/cal/east_asian_traditional/simple.rs
@@ -7,11 +7,11 @@ use calendrical_calculations::{gregorian::DAYS_IN_400_YEAR_CYCLE, rata_die::Rata
 
 macro_rules! day_fraction_to_ms {
     ($n:tt $(/ $d:tt)+) => {{
-        Milliseconds((MILLISECONDS_IN_EPHEMERIS_DAY as i128 * $n as i128 $( / $d as i128)+) as i64)
+        Milliseconds((MILLISECONDS_IN_EPHEMERIS_DAY as i128 * $n $( / $d)+) as i64)
     }};
     ($n:tt $(/ $d:tt)+, exact) => {{
         let d = day_fraction_to_ms!($n $(/ $d)+);
-        assert!((d.0 as i128 $(* $d as i128)+) % MILLISECONDS_IN_EPHEMERIS_DAY as i128 == 0, "inexact");
+        assert!((d.0 as i128 $(* $d)+) % MILLISECONDS_IN_EPHEMERIS_DAY as i128 == 0, "inexact");
         d
     }};
 }
@@ -23,16 +23,16 @@ pub(super) const BEIJING_UTC_OFFSET: Milliseconds = day_fraction_to_ms!(1397 / 1
 
 /// The mean year length according to the Gregorian solar cycle.
 const MEAN_GREGORIAN_YEAR_LENGTH: Milliseconds =
-    day_fraction_to_ms!(DAYS_IN_400_YEAR_CYCLE / 400, exact);
+    day_fraction_to_ms!((DAYS_IN_400_YEAR_CYCLE as i128) / 400, exact);
 
 /// The mean solar term length according to the Gregorian solar cycle
 const MEAN_GREGORIAN_SOLAR_TERM_LENGTH: Milliseconds =
-    day_fraction_to_ms!(DAYS_IN_400_YEAR_CYCLE / 400 / 12, exact);
+    day_fraction_to_ms!((DAYS_IN_400_YEAR_CYCLE as i128) / 400 / 12, exact);
 
 /// The mean synodic length on Jan 1 2000 according to the [Astronomical Almanac (1992)].
 ///
 /// [Astronomical Almanac (1992)]: https://archive.org/details/131123ExplanatorySupplementAstronomicalAlmanac/page/n302/mode/1up
-const MEAN_SYNODIC_MONTH_LENGTH: Milliseconds = day_fraction_to_ms!(295305888531 / 10000000000i64);
+const MEAN_SYNODIC_MONTH_LENGTH: Milliseconds = day_fraction_to_ms!(295305888531 / 10000000000);
 
 /// Number of milliseconds in a day.
 const MILLISECONDS_IN_EPHEMERIS_DAY: i64 = 24 * 60 * 60 * 1000;
